backend/requirements.txt
langchain==1.2.9
langchain-openai
tavily-python==0.4.0
python-dotenv==1.0.1
langchain-core
# langchain_community
langchain-tavily==0.2.17
fastapi[standard]
chromadb 
sentence-transformers